fixed bug of bug)

+++ recipes/default.rb
@@ -24,11 +24,13 @@ if value_for_platform({"ubuntu" => {"10.04" => true}, "default" => false})
   # fix buggy behaviour of resolvconf in ubuntu 10.04 (debian bug #642222)
   line = %q(\[ -f "$BASEFILE" \] \&\& RSLVCNFFILES="$BASEFILE)
   sfix = %q(\[ -f "$BASEFILE" \] \&\& RSLVCNFFILES="$RSLVCNFFILES\n$BASEFILE")
-  unless %x(grep '#{line}' /etc/resolvconf/update.d/libc).empty?
-    %x(cat /etc/resolvconf/update.d/libc | sed '/#{line}/{N;s/.*/#{sfix}/}' > /tmp/resolvconf_libc_642222.fix;
-        cat /tmp/resolvconf_libc_642222.fix > /etc/resolvconf/update.d/libc;
-        rm /tmp/resolvconf_libc_642222.fix
-      )
+  bash "fix_resolvconf_libc" do
+    code <<-EOH
+      cat /etc/resolvconf/update.d/libc | sed '/#{line}/{N;s/.*/#{sfix}/}' > /tmp/resolvconf_libc_642222.fix
+      cat /tmp/resolvconf_libc_642222.fix > /etc/resolvconf/update.d/libc
+      rm  /tmp/resolvconf_libc_642222.fix
+    EOH
+    not_if "grep '#{line}' /etc/resolvconf/update.d/libc"
   end
 end
